유저, 게시글 연결 완료

+++ project/src/main/java/com/example/project/model/post/entity/PostEntity.java
@@ -26,7 +26,7 @@ public class PostEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idx", nullable = false, unique = true)
-    private Long idx;
+    private Integer idx;
 
     @Column(name = "title", nullable = false)
     private String title;
